Rename workload_reader
